requirements.txt
streamlit
pypdf
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
InstructorEmbedding